Verify corrupt column in from_csv
